+++ Magic/src/main/java/com/elmakers/mine/bukkit/maps/MapController.java
@@ -294,13 +294,7 @@ public class MapController implements com.elmakers.mine.bukkit.api.maps.MapContr
 
     // This is copied from MagicController, which I'm still trying to keep out of this class. Shrug?
     public ItemStack createMap(int mapId) {
-        short durability = CompatibilityLib.isCurrentVersion() ? 0 : (short)mapId;
-        ItemStack mapItem = CompatibilityLib.getDeprecatedUtils().createItemStack(DefaultMaterials.getFilledMap(), 1, durability);
-        if (CompatibilityLib.isCurrentVersion()) {
-            mapItem = CompatibilityLib.getItemUtils().makeReal(mapItem);
-            CompatibilityLib.getNBTUtils().setInt(mapItem, "map", mapId);
-        }
-        return mapItem;
+        return CompatibilityLib.getInventoryUtils().createMap(DefaultMaterials.getFilledMap(), mapId);
     }
 
     /**