+++ Magic/CHANGELOG.md
@@ -3,7 +3,8 @@
 
 # 10.9.4
 
- - Fix the Capture spell not saving custom entity data
+ - Fix the Capture spell not saving custom entity data in 1.20.5+
+ - Fix recent versions of 1.21.1 being broken with a ton of skull-related spam (thanks, alekso56!)
 
 # 10.9.3