+++ Magic/CHANGELOG.md
@@ -1,6 +1,8 @@
 
 # CHANGELOG
 
+# 10.9.5
+
 # 10.9.4
 
  - Fix the Capture spell not saving custom entity data in 1.20.5+